+++ packages/ethereum-contracts/contracts/superfluid/SuperToken.sol
@@ -688,6 +688,18 @@ contract SuperToken is
         _transferFrom(account, spender, recipient, amount);
     }
 
+    function operationSend(
+        address spender,
+        address recipient,
+        uint256 amount,
+        bytes memory userData
+    ) 
+        external override
+        onlyHost
+    {
+        _send(msg.sender, spender, recipient, amount, userData, "", true);
+    }
+
     function operationUpgrade(address account, uint256 amount)
         external override
         onlyHost