+++ packages/sdk-core/CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
 ### Added
+- `SuperToken` class has `send` function
+- `batchCall` supports new `send` batch operation
 - Added `downgradeTo` function on `SuperToken` class
 - Added toga events to be part of query